Upgrade to Astro 5.x
