+++ package.json
@@ -12,13 +12,13 @@
     "prettier": "prettier --write ."
   },
   "dependencies": {
-    "@astrojs/check": "^0.7.0",
-    "@astrojs/markdown-remark": "^5.2.0",
-    "@astrojs/mdx": "^3.1.2",
-    "@astrojs/react": "^3.6.2",
-    "@astrojs/rss": "^4.0.7",
-    "@astrojs/sitemap": "^3.1.6",
-    "@astrojs/tailwind": "^5.1.0",
+    "@astrojs/check": "^0.9.4",
+    "@astrojs/markdown-remark": "^6.0.1",
+    "@astrojs/mdx": "^4.0.3",
+    "@astrojs/react": "^4.1.2",
+    "@astrojs/rss": "^4.0.11",
+    "@astrojs/sitemap": "^3.2.1",
+    "@astrojs/tailwind": "^5.1.4",
     "@hbsnow/rehype-sectionize": "^1.0.7",
     "@iconify-json/lucide": "^1.2.4",
     "@iconify-json/mingcute": "^1.2.1",
@@ -32,7 +32,7 @@
     "@types/react": "^18.3.5",
     "@types/react-dom": "^18.3.0",
     "@vercel/analytics": "^1.4.1",
-    "astro": "^4.11.3",
+    "astro": "^5.1.2",
     "astro-embed": "^0.7.4",
     "astro-icon": "^1.1.1",
     "bootstrap-icons": "^1.11.3",
