+++ src/components/Breadcrumbs.astro
@@ -17,6 +17,7 @@ export interface BreadcrumbItem {
 
 interface Props {
   items: BreadcrumbItem[]
+  class?: string
 }
 
 const { items } = Astro.props
