+++ src/components/ProjectCard.astro
@@ -19,7 +19,7 @@ const { project, linkDisabled, large } = Astro.props
   class="overflow-hidden rounded-xl border p-4 transition-colors duration-300 ease-in-out has-[a:hover]:bg-secondary/50"
 >
   <Link
-    href={`/projects/${project.slug}`}
+    href={`/projects/${project.id}`}
     class={cn('block', linkDisabled && 'pointer-events-none')}
   >
     <div class="flex flex-wrap gap-4 lg:flex-nowrap lg:items-start">
