+++ src/components/VodCard.astro
@@ -14,7 +14,7 @@ const { vod, featured = false } = Astro.props
   'relative aspect-video overflow-hidden rounded-lg',
 ]}>
   <a
-    href={`/vods/watch/${vod.slug}`}
+    href={`/vods/watch/${vod.id}`}
     class="block group"
   >
     <img
