+++ src/content.config.ts
@@ -1,7 +1,8 @@
+import { glob } from 'astro/loaders'
 import { defineCollection, z } from 'astro:content'
 
 const blog = defineCollection({
-  type: 'content',
+  loader: glob({ pattern: '**/*.{md,mdx}', base: './src/content/blog'}),
   schema: ({ image }) =>
     z.object({
       title: z
@@ -17,12 +18,7 @@ const blog = defineCollection({
           'Description should be 155 characters or less for optimal Open Graph display.',
         ),
       date: z.coerce.date(),
-      image: image()
-        .refine((img) => img.width === 1200 && img.height === 630, {
-          message:
-            'The image must be exactly 1200px × 630px for Open Graph requirements.',
-        })
-        .optional(),
+      image: image().optional(),
       tags: z.array(z.string()).optional(),
       authors: z.array(z.string()).optional(),
       draft: z.boolean().optional(),
@@ -30,7 +26,7 @@ const blog = defineCollection({
 })
 
 const authors = defineCollection({
-  type: 'content',
+  loader: glob({ pattern: '**/*.{md,mdx}', base: './src/content/authors' }),
   schema: z.object({
     name: z.string(),
     pronouns: z.string().optional(),
@@ -50,16 +46,12 @@ const authors = defineCollection({
 })
 
 const projects = defineCollection({
-  type: 'content',
+  loader: glob({ pattern: '**/*.{md,mdx}', base: './src/content/projects' }),
   schema: ({ image }) =>
     z.object({
       name: z.string(),
       description: z.string(),
-      image: image()
-        .refine((img) => img.width === 128 && img.height === 128, {
-          message: 'The image must be exactly 128 × 128px for consistency.',
-        })
-        .optional(),
+      image: image().optional(),
       link: z.string().url(),
       license: z.string().optional(),
       github_repo: z.string().optional(),
@@ -67,7 +59,7 @@ const projects = defineCollection({
 })
 
 const vods = defineCollection({
-  type: 'content',
+  loader: glob({ pattern: '**/*.{md,mdx}', base: './src/content/vods' }),
   schema: z.object({
     title: z.string(),
     streamDate: z.coerce.date(),
