+++ src/pages/rss.xml.ts
@@ -24,7 +24,7 @@ export async function GET(context: APIContext) {
         title: item.data.title,
         description: item.data.description,
         pubDate: item.data.date,
-        link: `/${item.collection}/${item.slug}/`,
+        link: `/${item.collection}/${item.id}/`,
       })),
     })
   } catch (error) {
